Working version of diffusion (with pytorch). Added structure for cylindrical convolutions (havent tested)
