+++ scripts/CaloDiffu.py
@@ -6,13 +6,14 @@ import torch.nn as nn
 from torch.autograd import Variable
 from torchinfo import summary
 from utils import *
+from models import *
 
 #https://huggingface.co/blog/annotated-diffusion
 
 
 class CaloDiffu(nn.Module):
     """Diffusion based generative model"""
-    def __init__(self, data_shape,num_batch,name='SGM',config=None):
+    def __init__(self, data_shape,num_batch,name='SGM',config=None, cylindrical = False):
         super(CaloDiffu, self).__init__()
         self._data_shape = data_shape
         self._num_batch = num_batch
@@ -51,13 +52,13 @@ class CaloDiffu(nn.Module):
         self.sqrt_alphas_cumprod = torch.sqrt(self.alphas_cumprod)
         self.sqrt_one_minus_alphas_cumprod = torch.sqrt(1. - self.alphas_cumprod)
 
-        #self.sqrt_recip_alphas = torch.reshape(torch.sqrt(1.0 / self.alphas), (-1,1))
-        #self.sqrt_alphas_cumprod = torch.reshape(torch.sqrt(self.alphas_cumprod), (-1,1))
-        #self.sqrt_one_minus_alphas_cumprod = torch.reshape(torch.sqrt(1. - self.alphas_cumprod), (-1,1))
-
         self.posterior_variance = self.betas * (1. - alphas_cumprod_prev) / (1. - self.alphas_cumprod)
 
-        self.noise_predictor = CondUnet(dim = 16, init_dim =16, out_dim = 1, dim_mults = (1,1,1), channels = 1)
+        layer_sizes = config['LAYER_SIZE_UNET']
+        cond_dim = config['COND_SIZE_UNET']
+
+        self.noise_predictor = CondUnet(cond_dim = cond_dim, out_dim = 1, channels = 1, layer_sizes = layer_sizes, cylindrical = cylindrical)
+
             
         calo_summary_shape = list(copy.copy(self._data_shape))
         calo_summary_shape.insert(0, 1)
@@ -104,7 +105,6 @@ class CaloDiffu(nn.Module):
         sqrt_one_minus_alphas_cumprod_t = extract(self.sqrt_one_minus_alphas_cumprod, t, x.shape)
         sqrt_recip_alphas_t = extract(self.sqrt_recip_alphas, t, x.shape)
 
-        print(x.shape, t.shape, betas_t.shape)
         #noise_pred = self.model.predict([x, t, cond], batch_size = batch_size)
         noise_pred = self.noise_predictor(x, cond, t)
         
@@ -150,7 +150,9 @@ class CaloDiffu(nn.Module):
 
         start = time.time()
         #for i in tqdm(reversed(range(0, num_steps)), desc='sampling loop time step', total=num_steps):
-        time_steps = range(0, num_steps)
+        time_steps = list(range(num_steps))
+        time_steps.reverse()
+
         for time_step in time_steps:      
             times = torch.full((gen_size,), time_step, device=device, dtype=torch.long)
             img = self.p_sample(img, cond, times)
