+++ scripts/configs/config_dataset2_diffu.json
@@ -14,12 +14,14 @@
     'NLAYERS':3,
     'EARLYSTOP':25,
     'LAYER_SIZE_AE':[32,64, 64,32],
-    'DIM_RED':[0,2, 0, 2],
+    'DIM_RED_AE':[0,2, 0, 2],
+    'LAYER_SIZE_UNET' : [16, 16, 32, 64],
+    'COND_SIZE_UNET' : 128,
     'KERNEL':[3,3,3],
     'STRIDE':[3,2,2],
     'ACT':'swish',
     'EMBED':128,
-    'CHECKPOINT_NAME':'dataset2_nov10_diffu',
+    'CHECKPOINT_NAME':'dataset2_nov11_cyl',
     'EMAX':100.,
     'EMIN':1.,
     'ECUT':0.0000151,
@@ -27,6 +29,7 @@
     'SNR':0.2,
     'logE':True,
     'NORMED':False,
+    'CYLINDRICAL': True,
     'MAXDEP':2,
     'SHOWERMAP': 'sqrt',
 }
