+++ scripts/utils.py
@@ -5,15 +5,9 @@ import numpy as np
 import matplotlib.pyplot as plt
 from matplotlib import gridspec
 import matplotlib.ticker as mtick
-from einops import rearrange
-
-
 import torch
 import torch.nn as nn
-import torch.nn.functional as f
 
-from inspect import isfunction
-from functools import partial
 
 def split_data_np(data, frac=0.8):
     np.random.shuffle(data)
@@ -378,24 +372,6 @@ def polar_to_cart(polar_data,nr=9,nalpha=16,nx=12,ny=12):
                 cart_img[binx,biny]+=polar_data[alpha,r]
     return cart_img
 
-#some pytorch modules & useful functions
-def exists(x):
-    return x is not None
-
-
-def default(val, d):
-    if exists(val):
-        return val
-    return d() if isfunction(d) else d
-
-class Residual(nn.Module):
-    def __init__(self, fn):
-        super().__init__()
-        self.fn = fn
-
-    def forward(self, x, *args, **kwargs):
-        return self.fn(x, *args, **kwargs) + x
-
 class EarlyStopper:
     def __init__(self, patience=1, min_delta=0):
         self.patience = patience
@@ -414,309 +390,6 @@ class EarlyStopper:
         return False
 
 
-
-def extract(a, t, x_shape):
-    batch_size = t.shape[0]
-    out = a.gather(-1, t.cpu())
-    return out.reshape(batch_size, *((1,) * (len(x_shape) - 1))).to(t.device)
-
-
-class SinusoidalPositionEmbeddings(nn.Module):
-    def __init__(self, dim):
-        super().__init__()
-        self.dim = dim
-
-    def forward(self, time):
-        device = time.device
-        half_dim = self.dim // 2
-        embeddings = np.log(10000) / (half_dim - 1)
-        embeddings = torch.exp(torch.arange(half_dim, device=device) * -embeddings)
-        embeddings = time[:, None] * embeddings[None, :]
-        embeddings = torch.cat((embeddings.sin(), embeddings.cos()), dim=-1)
-        return embeddings  
-
-
-class Block(nn.Module):
-    def __init__(self, dim, dim_out, groups = 8):
-        super().__init__()
-        self.proj = nn.Conv3d(dim, dim_out, kernel_size = 3, padding = 1)
-        self.norm = nn.GroupNorm(groups, dim_out)
-        self.act = nn.SiLU()
-
-    def forward(self, x, scale_shift = None):
-        x = self.proj(x)
-        x = self.norm(x)
-
-        if exists(scale_shift):
-            scale, shift = scale_shift
-            x = x * (scale + 1) + shift
-
-        x = self.act(x)
-        return x
-
-class ResnetBlock(nn.Module):
-    """https://arxiv.org/abs/1512.03385"""
-    
-    def __init__(self, dim, dim_out, *, cond_emb_dim=None, groups=8):
-        super().__init__()
-        self.mlp = (
-            nn.Sequential(nn.SiLU(), nn.Linear(cond_emb_dim, dim_out))
-            if exists(cond_emb_dim)
-            else None
-        )
-
-        self.block1 = Block(dim, dim_out, groups=groups)
-        self.block2 = Block(dim_out, dim_out, groups=groups)
-        self.res_conv = nn.Conv3d(dim, dim_out, kernel_size = 1) if dim != dim_out else nn.Identity()
-
-    def forward(self, x, time_emb=None):
-        h = self.block1(x)
-
-        if exists(self.mlp) and exists(time_emb):
-            time_emb = self.mlp(time_emb)
-            time_emb = rearrange(time_emb, "b c -> b c 1 1 1")
-            h = h + time_emb
-
-        h = self.block2(h)
-        return h + self.res_conv(x)
-
-class ConvNextBlock(nn.Module):
-    """https://arxiv.org/abs/2201.03545"""
-
-    def __init__(self, dim, dim_out, *, cond_emb_dim=None, mult=2, norm=True):
-        super().__init__()
-        self.mlp = (
-            nn.Sequential(nn.GELU(), nn.Linear(cond_emb_dim, dim))
-            if exists(cond_emb_dim)
-            else None
-        )
-
-        self.ds_conv = nn.Conv3d(dim, dim, kernel_size = 7, padding=3, groups=dim)
-
-        self.net = nn.Sequential(
-            nn.GroupNorm(1, dim) if norm else nn.Identity(),
-            nn.Conv3d(dim, dim_out * mult, kernel_size = 3, padding=1),
-            nn.GELU(),
-            nn.GroupNorm(1, dim_out * mult),
-            nn.Conv3d(dim_out * mult, dim_out, kernel_size =3, padding=1),
-        )
-
-        self.res_conv = nn.Conv3d(dim, dim_out, kernel_size = 1) if dim != dim_out else nn.Identity()
-
-    def forward(self, x, time_emb=None):
-        h = self.ds_conv(x)
-
-        if exists(self.mlp) and exists(time_emb):
-            condition = self.mlp(time_emb)
-            h = h + rearrange(condition, "b c -> b c 1 1")
-
-        h = self.net(h)
-        return h + self.res_conv(x)
-
-#TODO Fix to work in 3D
-class Attention(nn.Module):
-    def __init__(self, dim, heads=4, dim_head=32):
-        super().__init__()
-        self.scale = dim_head**-0.5
-        self.heads = heads
-        hidden_dim = dim_head * heads
-        self.to_qkv = nn.Conv2d(dim, hidden_dim * 3, 1, bias=False)
-        self.to_out = nn.Conv2d(hidden_dim, dim, 1)
-
-    def forward(self, x):
-        b, c, h, w = x.shape
-        qkv = self.to_qkv(x).chunk(3, dim=1)
-        q, k, v = map(
-            lambda t: rearrange(t, "b (h c) x y -> b h c (x y)", h=self.heads), qkv
-        )
-        q = q * self.scale
-
-        sim = einsum("b h d i, b h d j -> b h i j", q, k)
-        sim = sim - sim.amax(dim=-1, keepdim=True).detach()
-        attn = sim.softmax(dim=-1)
-
-        out = einsum("b h i j, b h d j -> b h i d", attn, v)
-        out = rearrange(out, "b h (x y) d -> b (h d) x y", x=h, y=w)
-        return self.to_out(out)
-
-#TODO Fix to work in 3D
-class LinearAttention(nn.Module):
-    def __init__(self, dim, heads=4, dim_head=32):
-        super().__init__()
-        self.scale = dim_head**-0.5
-        self.heads = heads
-        hidden_dim = dim_head * heads
-        self.to_qkv = nn.Conv2d(dim, hidden_dim * 3, 1, bias=False)
-
-        self.to_out = nn.Sequential(nn.Conv2d(hidden_dim, dim, 1), 
-                                    nn.GroupNorm(1, dim))
-
-    def forward(self, x):
-        b, c, h, w = x.shape
-        qkv = self.to_qkv(x).chunk(3, dim=1)
-        q, k, v = map(
-            lambda t: rearrange(t, "b (h c) x y -> b h c (x y)", h=self.heads), qkv
-        )
-
-        q = q.softmax(dim=-2)
-        k = k.softmax(dim=-1)
-
-        q = q * self.scale
-        context = torch.einsum("b h d n, b h e n -> b h d e", k, v)
-
-        out = torch.einsum("b h d e, b h d n -> b h e n", context, q)
-        out = rearrange(out, "b h c (x y) -> b (h c) x y", h=self.heads, x=h, y=w)
-        return self.to_out(out)
-
-class PreNorm(nn.Module):
-    def __init__(self, dim, fn):
-        super().__init__()
-        self.fn = fn
-        self.norm = nn.GroupNorm(1, dim)
-
-    def forward(self, x):
-        x = self.norm(x)
-        return self.fn(x)
-
-
-#downsample in two dimmensions but keep same dimm in first
-def Upsample(dim):
-    #return nn.ConvTranspose3d(dim, dim, kernel_size = (4,4,4), stride = (1,2,2), padding = (0,1,1))
-    return nn.Upsample(scale_factor = (1,2,2))
-
-def Downsample(dim):
-    #return nn.Conv3d(dim, dim, kernel_size = (4,4,4), stride = (1,2,2), padding = (0,1,1))
-    return nn.AvgPool3d(kernel_size = (1,2,2), stride = (1,2,2), padding =0)
-
-
-class CondUnet(nn.Module):
-#Unet with conditional layers
-    def __init__(
-        self,
-        dim,
-        init_dim=None,
-        out_dim=None,
-        dim_mults=(1, 2, 4, 8),
-        channels=1,
-        resnet_block_groups=8,
-        use_convnext=False,
-        use_attn = False,
-        convnext_mult=2,
-    ):
-        super().__init__()
-
-        # determine dimensions
-        self.channels = channels
-        self.use_attn = use_attn
-
-        #init_dim = default(init_dim, dim // 3 * 2)
-        #self.init_conv = nn.Conv3d(channels, init_dim, kernel_size =7, padding=3)
-
-        dims = [channels, *map(lambda m: dim * m, dim_mults)]
-        in_out = list(zip(dims[:-1], dims[1:])) 
-        
-        if use_convnext:
-            block_klass = partial(ConvNextBlock, mult=convnext_mult)
-        else:
-            block_klass = partial(ResnetBlock, groups=resnet_block_groups)
-
-        # time embeddings
-        time_dim = dim * 4
-
-        self.time_mlp = nn.Sequential(
-            SinusoidalPositionEmbeddings(dim),
-            nn.Linear(dim, time_dim),
-            nn.GELU(),
-            nn.Linear(time_dim, time_dim),
-        )
-
-        self.cond_mlp = nn.Sequential(
-            SinusoidalPositionEmbeddings(dim),
-            nn.Linear(dim, time_dim),
-            nn.GELU(),
-            nn.Linear(time_dim, time_dim),
-        )
-
-        cond_dim = time_dim*2
-
-        # layers
-        self.downs = nn.ModuleList([])
-        self.ups = nn.ModuleList([])
-        num_resolutions = len(in_out)
-
-
-
-        for ind, (dim_in, dim_out) in enumerate(in_out):
-            is_last = ind >= (num_resolutions - 1)
-
-            self.downs.append(
-                nn.ModuleList(
-                    [
-                        block_klass(dim_in, dim_out, cond_emb_dim=cond_dim),
-                        block_klass(dim_out, dim_out, cond_emb_dim=cond_dim),
-                        Residual(PreNorm(dim_out, LinearAttention(dim_out))),
-                        Downsample(dim_out) if not is_last else nn.Identity(),
-                    ]
-                )
-            )
-
-        mid_dim = dims[-1]
-        self.mid_block1 = block_klass(mid_dim, mid_dim, cond_emb_dim=cond_dim)
-        self.mid_attn = Residual(PreNorm(mid_dim, Attention(mid_dim)))
-        self.mid_block2 = block_klass(mid_dim, mid_dim, cond_emb_dim=cond_dim)
-
-        for ind, (dim_in, dim_out) in enumerate(reversed(in_out[1:])):
-            is_last = ind >= (num_resolutions - 1)
-
-            self.ups.append(
-                nn.ModuleList(
-                    [
-                        block_klass(dim_out * 2, dim_in, cond_emb_dim=cond_dim),
-                        block_klass(dim_in, dim_in, cond_emb_dim=cond_dim),
-                        Residual(PreNorm(dim_in, LinearAttention(dim_in))),
-                        Upsample(dim_in) if not is_last else nn.Identity(),
-                    ]
-                )
-            )
-
-        out_dim = default(out_dim, channels)
-        self.final_conv = nn.Sequential(
-            block_klass(dim, dim), nn.Conv3d(dim, out_dim, 1)
-        )
-
-    def forward(self, x, cond, time):
-
-        #x = self.init_conv(x)
-
-        t = self.time_mlp(time)
-        c = self.cond_mlp(cond)
-        conditions = torch.cat([t,c], axis = -1)
-
-
-        h = []
-
-        # downsample
-        for block1, block2, attn, downsample in self.downs:
-            x = block1(x, conditions)
-            x = block2(x, conditions)
-            if(self.use_attn): x = attn(x)
-            h.append(x)
-            x = downsample(x)
-
-        # bottleneck
-        x = self.mid_block1(x, conditions)
-        if(self.use_attn): x = self.mid_attn(x)
-        x = self.mid_block2(x, conditions)
-
-        # upsample
-        for block1, block2, attn, upsample in self.ups:
-            x = torch.cat((x, h.pop()), dim=1)
-            x = block1(x, conditions)
-            x = block2(x, conditions)
-            if(self.use_attn): x = attn(x)
-            x = upsample(x)
-
-        return self.final_conv(x)
 
 def SetFig(xlabel,ylabel):
     fig = plt.figure(figsize=(8, 6))
